refactor: extract principle scoring into _principle module
Split engine.py (271 lines) into engine.py (47 lines, public API) and
_principle.py (138 lines, principle-level scoring logic).
